+++ src/builtins/mod.rs
@@ -1689,6 +1689,8 @@ fn make_typed_array_intrinsic_in_env(vm: &mut Vm, env: GcIdx) -> error::Result<(
         vm.new_native_function_in_env("findIndex", typed_array_find_index, 1, env)?;
     let typed_array_find_last_fn =
         vm.new_native_function_in_env("findLast", typed_array_find_last, 1, env)?;
+    let typed_array_find_last_index_fn =
+        vm.new_native_function_in_env("findLastIndex", typed_array_find_last_index, 1, env)?;
     let typed_array_join_fn = vm.new_native_function_in_env("join", typed_array_join, 1, env)?;
     let typed_array_reverse_fn =
         vm.new_native_function_in_env("reverse", typed_array_reverse, 0, env)?;
@@ -1751,6 +1753,10 @@ fn make_typed_array_intrinsic_in_env(vm: &mut Vm, env: GcIdx) -> error::Result<(
             props.insert(
                 PropertyKey::from("findLast"),
                 data_prop(Value::Object(typed_array_find_last_fn)),
+            );
+            props.insert(
+                PropertyKey::from("findLastIndex"),
+                data_prop(Value::Object(typed_array_find_last_index_fn)),
             );
             props.insert(
                 PropertyKey::from("join"),